Embeddings require that input tensor be of dtype long.

+++ vocab.py
@@ -164,7 +164,7 @@ class VocabEntry(object):
 
         word_ids = self.words2charindices(sents)
         sents_padded = pad_sents_char(word_ids, self.char2id['<pad>'])
-        sents_padded_tensor = torch.tensor(sents_padded, dtype=torch.int, device=device)
+        sents_padded_tensor = torch.tensor(sents_padded, dtype=torch.long, device=device)
         sents_padded_tensor_reshaped = sents_padded_tensor.permute(1, 0, 2)
         return sents_padded_tensor_reshaped
         ### END YOUR CODE
